statistics/requirements.txt
kafka-python==2.0.2
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
textblob==0.17.1
faust-streaming==0.10.13